Move more headers to public
