+++ include/ccf/http_query.h
@@ -5,8 +5,8 @@
 #include "ds/nonstd.h"
 
 #define FMT_HEADER_ONLY
-#include <fmt/format.h>
 #include <charconv>
+#include <fmt/format.h>
 #include <map>
 #include <optional>
 #include <string_view>
