+++ samples/apps/nobuiltins/nobuiltins.cpp
@@ -3,11 +3,11 @@
 #include "ccf/app_interface.h"
 #include "ccf/base_endpoint_registry.h"
 #include "ccf/common_auth_policies.h"
+#include "ccf/json_handler.h"
 #include "ds/json.h"
 #include "enclave/node_context.h"
 #include "node/network_tables.h"
 #include "node/rpc/frontend.h"
-#include "node/rpc/json_handler.h"
 
 #include <charconv>
