+++ src/endpoints/common_endpoint_registry.cpp
@@ -6,11 +6,11 @@
 #include "ccf/common_auth_policies.h"
 #include "ccf/historical_queries_adapter.h"
 #include "ccf/http_query.h"
+#include "ccf/json_handler.h"
 #include "ds/nonstd.h"
 #include "enclave/node_context.h"
 #include "http/http_consts.h"
 #include "node/code_id.h"
-#include "node/rpc/json_handler.h"
 
 namespace ccf
 {
