+++ src/node/rpc/test/frontend_test.cpp
@@ -2,6 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT
 #include "ccf/app_interface.h"
+#include "ccf/json_handler.h"
 #include "ccf/user_frontend.h"
 #include "consensus/aft/request.h"
 #include "ds/files.h"
@@ -13,7 +14,6 @@
 #include "node/genesis_gen.h"
 #include "node/history.h"
 #include "node/network_state.h"
-#include "node/rpc/json_handler.h"
 #include "node/rpc/member_frontend.h"
 #include "node/rpc/node_frontend.h"
 #include "node/rpc/serdes.h"
